Require fname/lname, removed /users and /commutes content
A FB user with no fname or lname caused the users report to fail, so I
added those fields as required.

Also, New Relic reports CommutesController#index and
UsersController#index as one of the top 5 most time consuming
transactions, and we don't even link to those pages, so I changed the
views for those as "not allowed".  Not sure if that will help overall
performance, but no one needs these pages anyways.